Recompiled for the live site.
The last commit introduced an error message at the top of the webpage. This was probably caused by an error in compilation, which I did not see because the last time I tested it before commit, there was no error in the browser.

So here is the latest working compiled CSS.

With some more cleanup and Sass structuring.